working out

+++ src/lbgargs.py
@@ -59,7 +59,7 @@ _parser.add_option("--error",
 
 _parser.add_option("-c", "--compress",
                    dest="compress",
-                   type='long',
-                   default=1 / 2,
+                   type='float',
+                   default=1 / 2.,
                    help="Compress value",
                    )
